+++ src/utils/retry-handler.ts
@@ -1,25 +1,7 @@
 import { SendMessageCommand, SQSClient } from "@aws-sdk/client-sqs"
+import { getMaxRetryCount } from "../config/retry-config"
 import { GLOBAL } from "../singleton"
 import type { MeetingParams } from "../types"
-
-export const MAX_RETRY_COUNT = 2
-// Zoom web joins are probabilistic — the anti-bot / RTMS wall keys on the exit
-// IP's reputation, so each retry cycles onto a fresh residential IP. Give Zoom
-// more attempts than the default before giving up.
-export const ZOOM_WEB_MAX_RETRY_COUNT = 5
-
-/**
- * Max SQS retries for the current bot: higher for Zoom (web) so it can cycle
- * exit IPs past the anti-bot wall. Guarded so a crash before params are set
- * (GLOBAL unset) falls back to the default instead of throwing.
- */
-export function getMaxRetryCount(): number {
-  try {
-    return GLOBAL.get().meeting_platform === "zoom" ? ZOOM_WEB_MAX_RETRY_COUNT : MAX_RETRY_COUNT
-  } catch {
-    return MAX_RETRY_COUNT
-  }
-}
 
 /**
  * Creates SQS client with same credential logic as smart-rabbit
